level 1 + reading data from runs

+++ Assets/Scripts/DataSaving/SaveFileData.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 public class SaveFileData : ScriptableObject
 {
     public string fileName;
-    public int gameInProgress;    
+    public ulong gameInProgress;    
 }
